[PAGOPA-751] refactor: defined parameter for DB migration
The previous commits were defined in order to execute the DB migration with Flyway only if it is approved by user. This was defined using a dedicated task named ManualValidation. But if the user does not accepts the approval, the deploy cannot starts correctly and set its start also on DB migration 'Failed' state can cause misleading issues. Now, it will be used a dedicated parameters for disabling the migration if not needed.