[PAGOPA-975] fi: resolved bug on organization fiscal code set
With this commit, it was resolved a bug on which the organization fiscal code will always be set with the creditor institution code passed as path param. This bug did not permit a CI to set a specific value of the organization fiscal code to a transfer that must be transferred to another institution.